+++ src/main/java/com/lowdragmc/mbd2/common/data/MBDRecipeCapabilities.java
@@ -13,6 +13,7 @@ import com.lowdragmc.mbd2.integration.gtm.GTMEnergyRecipeCapability;
 import com.lowdragmc.mbd2.integration.mekanism.MekanismChemicalRecipeCapability;
 import com.lowdragmc.mbd2.integration.mekanism.MekanismHeatRecipeCapability;
 import com.lowdragmc.mbd2.integration.naturesaura.NaturesAuraRecipeCapability;
+import com.lowdragmc.mbd2.integration.pneumaticcraft.PNCPressureAirRecipeCapability;
 import net.minecraftforge.fml.ModLoader;
 
 public class MBDRecipeCapabilities {
@@ -42,6 +43,9 @@ public class MBDRecipeCapabilities {
         }
         if (MBD2.isNaturesAuraLoaded()) {
             MBDRegistries.RECIPE_CAPABILITIES.register(NaturesAuraRecipeCapability.CAP.name, NaturesAuraRecipeCapability.CAP);
+        }
+        if (MBD2.isPneumaticCraftLoaded()) {
+            MBDRegistries.RECIPE_CAPABILITIES.register(PNCPressureAirRecipeCapability.CAP.name, PNCPressureAirRecipeCapability.CAP);
         }
         ModLoader.get().postEvent(new MBDRegistryEvent.RecipeCapability());
         MBDRegistries.RECIPE_CAPABILITIES.freeze();