+++ src/main/java/com/lowdragmc/mbd2/common/graphprocessor/GraphParameterGet.java
@@ -11,6 +11,10 @@ import java.lang.annotation.Target;
 @Target({ElementType.FIELD})
 public @interface GraphParameterGet {
     /**
+     * The identity name of the parameter
+     */
+    String identity() default "";
+    /**
      * The display name of the parameter
      */
     String displayName() default "";