+++ src/main/java/com/lowdragmc/mbd2/common/gui/editor/machine/widget/TraitList.java
@@ -9,19 +9,16 @@ import com.lowdragmc.lowdraglib.gui.widget.DraggableScrollableWidgetGroup;
 import com.lowdragmc.lowdraglib.gui.widget.ImageWidget;
 import com.lowdragmc.lowdraglib.gui.widget.SelectableWidgetGroup;
 import com.lowdragmc.lowdraglib.utils.Size;
-import com.lowdragmc.mbd2.MBD2;
 import com.lowdragmc.mbd2.api.registry.MBDRegistries;
 import com.lowdragmc.mbd2.common.gui.editor.MachineEditor;
 import com.lowdragmc.mbd2.common.gui.editor.MachineProject;
 import com.lowdragmc.mbd2.common.gui.editor.machine.MachineScenePanel;
 import com.lowdragmc.mbd2.common.trait.TraitDefinition;
-import com.lowdragmc.mbd2.integration.ldlib.MBDLDLibPlugin;
 import lombok.Getter;
 import net.minecraftforge.api.distmarker.Dist;
 import net.minecraftforge.api.distmarker.OnlyIn;
 
 import javax.annotation.Nullable;
-import java.lang.reflect.InvocationTargetException;
 
 public class TraitList extends DraggableScrollableWidgetGroup {
     private final MachineEditor editor;
@@ -85,7 +82,7 @@ public class TraitList extends DraggableScrollableWidgetGroup {
             var menu = TreeBuilder.Menu.start()
                     .branch(Icons.ADD_FILE, "editor.machine.machine_traits.add_trait", m -> {
                         for (var wrapper : MBDRegistries.TRAIT_DEFINITIONS.values()) {
-                            m.leaf("config.definition.trait.%s.name".formatted(wrapper.annotation().name()), () -> {
+                            m.leaf(wrapper.creator().get().getIcon(), "config.definition.trait.%s.name".formatted(wrapper.annotation().name()), () -> {
                                 var traitDefinition = wrapper.creator().get();
                                 if (!traitDefinition.allowMultiple()) {
                                     for (var existed : project.getDefinition().machineSettings().traitDefinitions()) {