+++ src/main/java/com/lowdragmc/mbd2/common/machine/definition/config/event/MachineCustomDataUpdateEvent.java
@@ -1,10 +1,14 @@
 package com.lowdragmc.mbd2.common.machine.definition.config.event;
 
 import com.lowdragmc.lowdraglib.gui.editor.annotation.LDLRegister;
+import com.lowdragmc.lowdraglib.gui.graphprocessor.data.parameter.ExposedParameter;
 import com.lowdragmc.mbd2.common.machine.MBDMachine;
 import com.lowdragmc.mbd2.common.graphprocessor.GraphParameterGet;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraftforge.eventbus.api.Cancelable;
+
+import java.util.Map;
+import java.util.Optional;
 
 @Cancelable
 @LDLRegister(name = "MachineCustomDataUpdateEvent", group = "MachineEvent")
@@ -18,5 +22,12 @@ public class MachineCustomDataUpdateEvent extends MachineEvent {
         super(machine);
         this.newValue = newValue;
         this.oldValue = oldValue;
+    }
+
+    @Override
+    public void bindParameters(Map<String, ExposedParameter> exposedParameters) {
+        super.bindParameters(exposedParameters);
+        Optional.ofNullable(exposedParameters.get("oldValue")).ifPresent(p -> p.setValue(oldValue));
+        Optional.ofNullable(exposedParameters.get("newValue")).ifPresent(p -> p.setValue(newValue));
     }
 }