+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # ChangeLog
 
 * add ja_jp translation (thanks to @A8T1N)
-* add Nature's Aura compat
-* 
+* add Nature's Aura - aura
+* add Pneumatic Craft  - pressure / air
+* fix event node graph doesn't work in some cases
 * fix kjs recipe functions